revised auth

+++ user/const.go
@@ -16,7 +16,7 @@ const UserFileTable = `CREATE TABLE IF NOT EXISTS user_files (
 const SigningKeyLength uint8 = 32
 
 type User struct {
-	ID       uint64 `json:"id"`
+	ID       uint32 `json:"id"`
 	Name     string `json:"name"`
 	Email    string `json:"email"`
 	Password string `json:"password"`
